Move types to types/config.go
